drunk update2
